changed cities

+++ routes/api.php
@@ -1,36 +1,35 @@
 <?php
 
-use App\Constants\DB\EventDBConstants;
-use App\Constants\Request\EventRequestConstants;
 use App\Http\Controllers\ApplierController;
+use App\Http\Controllers\ChatController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\EventController;
 use App\Http\Controllers\CommentController;
-use App\Http\Controllers\ManufacturerController;
-use App\Http\Controllers\ColorController;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoryController;
-use App\Http\Controllers\ChatController;
 use App\Http\Controllers\CityController;
+use App\Http\Controllers\CommunityController;
 use App\Http\Controllers\ComplaintController;
 use App\Http\Controllers\CountryController;
 use App\Http\Controllers\EventArchiveController;
 use App\Http\Controllers\InteresterController;
 use App\Http\Controllers\MediaController;
 use App\Http\Controllers\MessageController;
+use App\Http\Controllers\PlaceController;
 use App\Http\Controllers\QuestionController;
+use App\Http\Controllers\RegionController;
 use App\Http\Controllers\TagController;
 use App\Http\Controllers\TestController;
 use App\Http\Controllers\UserController;
 use App\Http\Middleware\AdminIsAuthorized;
-use App\Http\Middleware\CheckAuthor;
+use App\Http\Middleware\CheckChatMember;
 use App\Http\Middleware\CheckCommentAuthor;
 use App\Http\Middleware\CheckEventAuthor;
+use App\Http\Middleware\CheckEventDoNotAuthor;
+use App\Http\Middleware\CheckMessageAuthor;
 use App\Http\Middleware\CheckQuestionAuthor;
 use App\Http\Middleware\UserIsAuthorized;
-use App\Models\Event;
-use Illuminate\Database\Eloquent\Builder;
 use Laravel\Scout\Scout;
 
 /*
@@ -50,30 +49,44 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::get('events', [EventController::class, 'index']);
 Route::get('events/search', [EventController::class, 'searchEvent']);
-Route::get('events/addAppliers', [EventController::class, 'addAppliers']);
-Route::post('events/{id}/photos/upload', [EventController::class, 'addPhotos']);
-Route::patch('events/{id}/photos/delete', [EventController::class, 'deletePhotos']);
-Route::patch('events/{id}/photos/deleteMain', [EventController::class, 'deleteMainPhoto']);
+Route::post('events/{id}/photos/upload', [EventController::class, 'addPhotos'])->middleware([UserIsAuthorized::class]);
+Route::patch('events/{id}/photos/delete', [EventController::class, 'deletePhotos'])->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
+Route::patch('events/{id}/photos/deleteMain', [EventController::class, 'deleteMainPhoto'])->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
 Route::post('events', [EventController::class, 'create'])->middleware([UserIsAuthorized::class]);
 Route::get('events/filter', [EventController::class, 'filter']);
 Route::put('events/{id}', [EventController::class, 'update'])->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
 Route::get('events/{id}/similar', [EventController::class, 'similar']);
 Route::delete('events/{id}', [EventController::class, 'delete'])->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
 Route::get('events/maili/search', [EventController::class, 'mailisearch']);
+Route::get('users/{id}/events', [EventController::class, 'getEventsByAuthorId']);
+
 Route::get('events/{id}', [EventController::class, 'show']);
 //eventArchives
 Route::get('eventArchives', [EventArchiveController::class, 'index']);
+Route::get('users/{id}/eventArchives', [EventArchiveController::class, 'showUserEventArchives']);
 Route::post('events/{id}/eventArchives', [EventArchiveController::class, 'archive'])->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
 Route::post('eventArchives/{id}/unarchive', [EventArchiveController::class, 'unarchive'])->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
 Route::delete('eventArchives/{id}', [EventArchiveController::class, 'delete'])->middleware([UserIsAuthorized::class]);
 Route::get('eventArchives/{id}', [EventArchiveController::class, 'show']);
 
-//cities
-Route::get('cities', [CityController::class, 'index']);
-Route::get('cities/{id}', [CityController::class, 'show']);
-Route::post('countries/{id}/cities', [CityController::class, 'create'])->middleware([AdminIsAuthorized::class]);
-Route::put('cities/{id}', [CityController::class, 'update'])->middleware([AdminIsAuthorized::class]);
-Route::delete('cities/{id}', [CityController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
+//places
+Route::get('communities/{communityId}/places', [PlaceController::class, 'CommunityPlaces']);
+Route::get('places/{id}', [PlaceController::class, 'show']);
+Route::delete('places/{id}', [PlaceController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
+Route::post('communities/{communityId}/places', [PlaceController::class, 'create'])->middleware([AdminIsAuthorized::class]);
+Route::put('communities/{communityId}/places/{id}', [PlaceController::class, 'update'])->middleware([AdminIsAuthorized::class]);
+//communities
+Route::get('regions/{regionsId}/communities', [CommunityController::class, 'RegionCommunities']);
+Route::get('communities/{id}', [CommunityController::class, 'show']);
+Route::delete('communities/{id}', [CommunityController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
+Route::post('regions/{regionsId}/communities', [CommunityController::class, 'create'])->middleware([AdminIsAuthorized::class]);
+Route::put('regions/{regionsId}/communities/{id}', [CommunityController::class, 'update'])->middleware([AdminIsAuthorized::class]);
+//regions
+Route::get('countries/{countryId}/regions', [RegionController::class, 'CountryRegions']);
+Route::get('regions/{id}', [RegionController::class, 'show']);
+Route::delete('regions/{id}', [RegionController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
+Route::post('countries/{countryId}/regions', [RegionController::class, 'create'])->middleware([AdminIsAuthorized::class]);
+Route::put('countries/{countryId}/regions/{id}', [RegionController::class, 'update'])->middleware([AdminIsAuthorized::class]);
 //countries
 Route::get('countries', [CountryController::class, 'index']);
 Route::get('countries/{id}', [CountryController::class, 'show']);
@@ -83,8 +96,10 @@ Route::delete('countries/{id}', [CountryController::class, 'delete'])->middlewar
 
 //comments
 Route::get('comments', [CommentController::class, 'index']);
+Route::get('users/{id}/comments', [CommentController::class, 'getCommentsByAuthorId']);
 Route::get('comments/{id}', [CommentController::class, 'show']);
 Route::post('events/{id}/comments', [CommentController::class, 'create'])->middleware([UserIsAuthorized::class]);
+Route::get('events/{id}/comments', [CommentController::class, 'getEventComments']);
 Route::put('comments/{id}', [CommentController::class, 'update'])->middleware([UserIsAuthorized::class, CheckCommentAuthor::class]);
 Route::delete('comments/{id}', [CommentController::class, 'delete'])->middleware([UserIsAuthorized::class, CheckCommentAuthor::class]);
 
@@ -102,16 +117,16 @@ Route::put('tags/{id}', [TagController::class, 'update'])->middleware([AdminIsAu
 Route::delete('tags/{id}', [TagController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
 //messages
 Route::get('messages', [MessageController::class, 'index']);
+// ->middleware([AdminIsAuthorized::class]);
 Route::get('messages/{id}', [MessageController::class, 'show']);
-Route::patch('messages/{id}', [MessageController::class, 'update']);
-Route::post('events/{eventId}/messages', [MessageController::class, 'create']);
+// ->middleware([AdminIsAuthorized::class]);
+Route::patch('messages/{id}', [MessageController::class, 'update'])->middleware([UserIsAuthorized::class, CheckMessageAuthor::class]);
+Route::post('events/{eventId}/messages', [MessageController::class, 'create'])->middleware([UserIsAuthorized::class]);
 //chats
-Route::get('chats', [ChatController::class, 'index']);
-Route::get('chats/author', [ChatController::class, 'getAllAuthorChat']);
-Route::get('chats/member', [ChatController::class, 'getAllMemberChat']);
-Route::get('chats/{id}', [ChatController::class, 'show']);
-Route::get('chats/{id}/messages', [ChatController::class, 'getChatWithAllMessages']);
-Route::delete('chats/{id}', [ChatController::class, 'delete']);
+Route::get('chats/author', [ChatController::class, 'getAllAuthorChat'])->middleware([UserIsAuthorized::class]);
+Route::get('chats/member', [ChatController::class, 'getAllMemberChat'])->middleware([UserIsAuthorized::class]);
+Route::get('chats/{id}', [ChatController::class, 'show'])->middleware([UserIsAuthorized::class], [CheckChatMember::class]);
+Route::get('chats/{id}/messages', [ChatController::class, 'getChatWithAllMessages'])->middleware([UserIsAuthorized::class, CheckChatMember::class]);
 //auth
 Route::post('login', [AuthController::class, 'login']);
 Route::post('register', [AuthController::class, 'register']);
@@ -121,16 +136,21 @@ Route::post('test', [TestController::class, 'maxDifferentWithDumbCondition']);
 
 //media
 
-Route::get('media', [MediaController::class, 'index']);
+Route::get('comments/{id}/media', [MediaController::class, 'getCommentMedia']);
+Route::get('events/{id}/media', [MediaController::class, 'getEventMedia']);
 Route::get('media/{id}', [MediaController::class, 'show']);
 Route::post('media/comments/{comment_id}', [MediaController::class, 'addPhotos']);
-// ->middleware([AdminIsAuthorized::class]);
+// ->middleware([UserIsAuthorized::class,CheckCommentAuthor::class]);
 Route::delete('media/{id}', [MediaController::class, 'delete']);
-// ->middleware([AdminIsAuthorized::class]);
+// ->middleware([AdminIsAuthorized::class,CheckMediaAuthor::class]);
 //questions
-Route::get('questions', [QuestionController::class, 'index']);
+Route::get('users/{id}/questions', [QuestionController::class, 'getQuestionsByAuthorId']);
+Route::get('events/{id}/questions', [QuestionController::class, 'getEventQuestions']);
 Route::get('questions/{id}', [QuestionController::class, 'show']);
-Route::post('events/{id}/questions', [QuestionController::class, 'create'])->middleware([UserIsAuthorized::class]);
+Route::post('events/{id}/questions/ask', [QuestionController::class, 'createQuestion'])
+    ->middleware([UserIsAuthorized::class, CheckEventDoNotAuthor::class]);
+Route::post('events/{id}/questions/response', [QuestionController::class, 'responseToQuestion'])
+    ->middleware([UserIsAuthorized::class, CheckEventAuthor::class]);
 Route::put('questions/{id}', [QuestionController::class, 'update'])->middleware([
     UserIsAuthorized::class
     // , CheckQuestionAuthor::class
@@ -140,17 +160,21 @@ Route::delete('questions/{id}', [QuestionController::class, 'delete'])->middlewa
     // , CheckQuestionAuthor::class
 ]);
 //appliers
-Route::get('appliers', [ApplierController::class, 'index']);
-Route::get('appliers/{id}', [ApplierController::class, 'show']);
-Route::put('events/{id}/appliers', [ApplierController::class, 'update'])->middleware([UserIsAuthorized::class]);
+Route::get('events/{id}/appliers', [ApplierController::class, 'EventAppliers']);
+// ->middleware([AdminIsAuthorized::class]);
+Route::get('events/{id}/appliers/count', [ApplierController::class, 'applierCount']);
+Route::get('appliers/{id}', [ApplierController::class, 'show'])->middleware([AdminIsAuthorized::class]);
+Route::put('events/{id}/appliers', [ApplierController::class, 'changeApplierStatus'])->middleware([UserIsAuthorized::class]);
 //interesters
-Route::get('interesters', [InteresterController::class, 'index']);
-Route::get('interesters/{id}', [InteresterController::class, 'show']);
-Route::post('interesters', [InteresterController::class, 'create'])->middleware([AdminIsAuthorized::class]);
-Route::put('interesters/{id}', [InteresterController::class, 'update'])->middleware([AdminIsAuthorized::class]);
-Route::delete('interesters/{id}', [InteresterController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
+Route::get('events/{id}/interesters', [InteresterController::class, 'EventInteresters']);
+// ->middleware([AdminIsAuthorized::class]);
+Route::get('events/{id}/interesters/count', [InteresterController::class, 'InteresterCount']);
+Route::get('interesters/{id}', [InteresterController::class, 'show'])->middleware([AdminIsAuthorized::class]);
+Route::put('events/{id}/interesters', [InteresterController::class, 'changeInteresterStatus'])->middleware([UserIsAuthorized::class]);
 //complaints complaints/018dc206-75b3-7162-8a0e-1bb416cda147/unassign
 Route::get('complaints', [ComplaintController::class, 'index']);
+Route::get('events/{id}/authorComplaints', [ComplaintController::class, 'getAuthorComplaints']);
+Route::get('events/{id}/receiverComplaints', [ComplaintController::class, 'getReceiverComplaints']);
 Route::get('complaints/filter', [ComplaintController::class, 'filter']);
 Route::get('complaints/{id}', [ComplaintController::class, 'show']);
 Route::post('events/{id}/complaints', [ComplaintController::class, 'create'])->middleware([AdminIsAuthorized::class]);
@@ -164,16 +188,12 @@ Route::delete('complaints/{id}', [ComplaintController::class, 'delete'])->middle
 Route::get('users', [UserController::class, 'index']);
 // ->middleware([AdminIsAuthorized::class]);
 Route::get('users/{id}', [UserController::class, 'show']);
-Route::post('users/{id}/avatars', [UserController::class, 'addPhoto']);
-Route::delete('users/{id}/avatars', [UserController::class, 'deletePhoto']);
-Route::get('users/{id}/events', [UserController::class, 'userEvents']);
-Route::get('users/{id}/questions', [UserController::class, 'userQuestions']);
-Route::get('users/{id}/comments', [UserController::class, 'userComments']);
 // ->middleware([AdminIsAuthorized::class]);
+Route::post('users/{id}/avatars', [UserController::class, 'addPhoto'])->middleware([AdminIsAuthorized::class]);
+Route::delete('users/{id}/avatars', [UserController::class, 'deletePhoto'])->middleware([AdminIsAuthorized::class]);
 Route::post('users', [UserController::class, 'create'])->middleware([AdminIsAuthorized::class]);
 Route::patch('users/{id}/banned', [UserController::class, 'banned'])->middleware([AdminIsAuthorized::class]);
-Route::put('users/{id}', [UserController::class, 'update']);
-// ->middleware([AdminIsAuthorized::class]);
+Route::put('users/{id}', [UserController::class, 'update'])->middleware([AdminIsAuthorized::class]);
 Route::delete('users/{id}', [UserController::class, 'delete'])->middleware([AdminIsAuthorized::class]);
